Create findLeaderInArray.cpp
Given an array A of positive integers. Your task is to find the leaders in the array. An element of array is leader if it is greater than or equal to all the elements to its right side. The rightmost element is always a leader.
Time Complexity: o(N) 
Space complexity:o(1)